Create Moving VM out from DS
Anyone want  to move all the VM from one Datastore to some other Datastore dynamically according to the free space present, in the same time it will move only the disk which are preset in the same/source Datastore. we just need to give the source DS name ESX name and vcenter details. Please find the script below